requirements.txt
numpy>=1.24.0
scipy>=1.10.0
rich>=13.0.0
matplotlib>=3.7.0
# GPU Support for NVIDIA A5000 (CUDA 12.x)
cupy-cuda12x>=12.0.0
